# swagger.json
{
  "info": {
    "title": "Lab Helpers",
    "version": "1.11.0",
    "description": "API que acompaña a Ansible Tower para levantar ambientes de desarrollo durante los workshops de DevOps 101 de Conatel, y otras instancias."
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json",
    "text/html"
  ],
  "host": "127.0.0.1:8081",
  "basePath": "",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "SecretAuthentication": {
      "type": "apiKey",
      "in": "header",
      "name": "X-Lab-Helpers-Secret"
    },
    "JWTTokenAuthentication": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "swagger": "2.0",
  "paths": {
...
    "/api/v1/files/{filename}": {
      "get": {
        "summary": "Devuelve el archivo almacenado en el servidor según su `filename`",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "type": "string",
            "minimum": 1,
            "maximum": 1,
            "description": "Nombre del archivo almacenado."
          }
        ],
        "security": [
          {
            "JWTTokenAuthentication": []
          }
        ],
        "tags": [
          "files"
        ],
        "responses": {
          "200": {
            "description": "El archivo buscado.",
            "schema": {
              "type": "file"
            }
          },
          "400": {
            "$ref": "#/responses/Error"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
    "/api/auth/": {
      "get": {
        "summary": "Devuelve información básica de la API",
        "description": "Se puede utilizar este endpoint para recorrer la API de autenticación. En la respuesta se detallan todos los endpoints disponibles bajo este namespace.\n",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "login": "/helpers/api/auth/login/",
                "signup": "/helpers/api/auth/signup/"
              }
            }
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
...
    "/api/v1/documents/": {
      "get": {
        "summary": "Devuelve una lista de documentos paginada.",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "type": "number",
            "minimum": 1,
            "maximum": 100,
            "description": "Limite de documentos a devolver."
          },
          {
            "in": "query",
            "name": "offset",
            "type": "number",
            "minimum": 0,
            "description": "Valor de offset según el cual se paginan los resultados."
          },
          {
            "in": "query",
            "name": "sortBy",
            "type": "string",
            "default": "id",
            "description": "Valor de la variable por la cual se ordenan los resultados.\n"
          }
        ],
        "security": [
          {
            "JWTTokenAuthentication": []
          }
        ],
        "tags": [
          "documents"
        ],
        "responses": {
          "200": {
            "description": "La lista de documentos paginada.",
            "schema": {
              "$ref": "#/definitions/Documents"
            }
          },
          "400": {
            "$ref": "#/responses/Error"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      },
      "post": {
        "summary": "Almacena un documento.",
        "description": "El documento a almacenar puede ser cualquier objeto JSON. La aplicación almacenara el documento en la base, y devolverá el ID donde puede recuperarse.\n",
        "parameters": [
          {
            "in": "body",
            "required": true,
            "name": "document",
            "description": "Objeto JSON.",
            "schema": {
              "$ref": "#/definitions/AnyObject"
            }
          }
        ],
        "security": [
          {
            "JWTTokenAuthentication": []
          }
        ],
        "tags": [
          "documents"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Document"
            }
          },
          "400": {
            "$ref": "#/responses/Error"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AnyObject": {
      "description": "Cualquier objeto.",
      "type": "object",
      "example": {
        "something": "awesome"
      }
    },
    "ID": {
      "description": "ID del recurso. Compatible con CUID.",
      "type": "string",
      "pattern": "^c[^\\s-]{8,}$",
      "example": "cjld2cjxh0000qzrmn831i7rn"
    },
    "Count": {
      "description": "Cantidad de elementos devueltos.",
      "type": "number",
      "example": 1
    },
    "DocumentURL": {
      "description": "URL correspondiente al recurso.",
      "type": "string",
      "pattern": "^(/[^/ ]*)+/?$",
      "example": "/helpers/api/v1/documents/cjld2cjxh0000qzrmn831i7rn"
    },
...
    "CreatedDate": {
      "description": "Fecha de creación del recuro",
      "type": "string",
      "pattern": "\\d{4}-[01]\\d-[0-3]\\dT[0-2]\\d:[0-5]\\d:[0-5]\\d\\.\\d+",
      "example": "2018-10-19T14:06:44.271Z"
    },
    "UpdatedDate": {
      "description": "Fecha de actualización del recuro",
      "type": "string",
      "pattern": "\\d{4}-[01]\\d-[0-3]\\dT[0-2]\\d:[0-5]\\d:[0-5]\\d\\.\\d+",
      "example": "2018-10-20T15:26:44.271Z"
    },
    "Error": {
      "description": "Objeto con mensaje de error.",
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "description": "Mensaje de error.",
          "example": "no existe un recurso identificado por la id 'cjld2cjxh0000qzrmn831i7rn'"
        }
      }
    },
...
    "Next": {
      "description": "URL para obtener la siguiente lista de recursos, paginados utilizando las opciones `offset` y `limit`.",
      "type": "string",
      "pattern": "^(/[^/ ]*)+/?$",
      "example": "/helpers/api/v1/documents/?offset=20&limit=10"
    },
    "Previous": {
      "description": "URL para obtener la anterior lista de recursos, paginados utilizando las opciones `offset` y `limit`.",
      "type": "string",
      "pattern": "^(/[^/ ]*)+/?$",
      "example": "/helpers/api/v1/documents/?offset=0&limit=10"
    },
    "Documents": {
      "type": "object",
      "description": "Lista de documentos.",
      "properties": {
        "count": {
          "$ref": "#/definitions/Count"
        },
        "next": {
          "$ref": "#/definitions/Next"
        },
        "previous": {
          "$ref": "#/definitions/Previous"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Document"
          }
        }
      }
    },
...
    "Document": {
      "type": "object",
      "description": "Documento JSON almacenado en la base.",
      "properties": {
        "created": {
          "$ref": "#/definitions/CreatedDate"
        },
        "data": {
          "$ref": "#/definitions/AnyObject"
        },
        "id": {
          "$ref": "#/definitions/ID"
        },
        "type": {
          "type": "string",
          "enum": [
            "documents"
          ],
          "description": "Tipo del recurso.",
          "example": "document"
        },
        "url": {
          "$ref": "#/definitions/DocumentURL"
        },
        "updated": {
          "$ref": "#/definitions/UpdatedDate"
        }
      }
    }
  },
  "responses": {
    "Unauthorized": {
      "description": "No cuenta con credenciales suficientes para acceder a este recurso.\n",
      "schema": {
        "type": "string",
        "example": "Unauthorized"
      }
    },
    "NotFound": {
      "description": "El recurso requerido, un documento por ejemplo, no fue encontrado.\n",
      "schema": {
        "type": "string",
        "example": "Not Found"
      }
    },
    "ServerError": {
      "description": "Algo sucedio en el servidor y no puede responder\n",
      "schema": {
        "type": "string",
        "example": "Server Error"
      }
    },
...
    "Error": {
      "description": "El request es invalido o no puede procesarse. Un mensaje de error con más información seré provisto en el cuerpo del mensaje.\n",
      "schema": {
        "$ref": "#/definitions/Error"
      }
    }
  },
...
}